% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caschool}
\alias{caschool}
\title{THE CALIFORNIA TEST SCORE DATA SET}
\format{
\describe{
\item{observation_n~r}{Observation number}
\item{dist_cod}{district code}
\item{county}{County name}
\item{district}{District name}
\item{gr_span}{}
...
}
}
\source{
\url{www.cde.ca.gov}
}
\usage{
caschool
}
\description{
This dataset contains data on test performance,
school characteristics and student demographic backgrounds. The data used here are from all 420 K-6 and K-8 districts in
California with data available for 1998 and 1999
}
\keyword{datasets}
