\name{Cig8595}
\docType{data}
\alias{Cig8595}
\title{The cigarette consumption panel data set}
\description{
  a panel of 48 observations  from 1985  to 1995

 \emph{number of observations} :  528

 \emph{observation} :  region

   \emph{country} :  United States

}
\usage{data(Cig8595)}
\format{A dataframe containing :
  \describe{
    \item{state}{state}
    \item{year}{year}
    \item{cpi}{consummer price index}
    \item{pop}{state population}
    \item{packpc}{number of packs per capita}
    \item{income}{state personal income (total, nominal)}
    \item{tax}{average state, federal, and average local excise taxes for fiscal year}
    \item{avgprs}{average price during fiscal year, including sales taxes}
    \item{taxs}{average excise taxes for fiscal year, including sales taxes}
  }
}
\source{
    Professor Jonhatan Gruber, MIT.
}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers.
}
\keyword{datasets}