% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cig_ch10}
\alias{cig_ch10}
\title{THE CIGARETTE CONSUMPTION PANEL DATA SET}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{state}{US State}
\item{year}{Year}
\item{cpi}{Consumer price index}
\item{pop}{Population}
\item{income}{state personal income}
\item{tax}{average state, federal, and average local excise taxes for fiscal year}
\item{avgprs}{average price during fiscal year}
\item{taxs}{average excise taxes for fiscal year}
...
}
}
\usage{
cig_ch10
}
\description{
The data set consists of annual data for the 48 continental U.S. states from 1985 Ð 1995.
Quantity consumed is measured by annual per capita cigarette sales in packs per fiscal year,
as derived from state tax collection data.  The price is the average retail cigarette price
per pack during the fiscal year, including taxes.  Income is per capita income.
The general sales tax is the average tax, in cents per pack, due to the broad-based state sales
tax applied to all consumption goods.  The cigarette-specific tax is the tax applied to
cigarettes only.  All prices, income, and taxes used in the regressions in this chapter
are deflated by the Consumer Price Index and thus are in constant (real) dollars.
We are grateful to Professor Jonathan Gruber of MIT for providing us with these data.

This data file only contains data from the year 1985 and 1995.
}
\keyword{datasets}
