\name{CPSch3}
\docType{data}
\alias{CPSch3}
\title{Earnings from the Current Population Survey}
\description{
  a cross-section from 1998

 \emph{number of observations} :  11130

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(CPSch3)}
\format{A dataframe containing :
  \describe{
    \item{sex}{1 if male, 2 if female}
    \item{year}{survey year}
    \item{ahe98}{average hourly earnings in \$1998}
  }
}
\source{
    Bureau of labor statistics, U.S. Departement of Labor.
}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers.
}
\keyword{datasets}