\name{Oj}
\docType{data}
\alias{Oj}
\title{The orange juice data set}
\description{
  monthly observations from 1948-01 to 2001-06

 \emph{number of observations} :  642

 \emph{observation} :  Sector

   \emph{country} :  United States

}
\usage{data(Oj)}
\format{A dataframe containing :
  \describe{
    \item{ENTRY}{year:quarter}
    \item{PPIOJ}{producer price for frozen orange juice}
    \item{PWFSA}{producer price index for finished goods}
    \item{FDD}{freezing degree days (from daily minimimum temperature reacorded at Orlando area airports)}
  }
}
\source{
    U.S. Bureau of Labor Statistics for PPIOJ and PWFSA, National Oceanic and Atmospheric Administration (NOAA) of the U.S Department of Commerce for FDD.
}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers.
}
\keyword{datasets}