% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caschool}
\alias{caschool}
\title{THE CALIFORNIA TEST SCORE DATA SET
This dataset contains data on test performance,
school characteristics and student demographic backgrounds. The data used here are from all 420 K-6 and K-8 districts in
California with data available for 1998 and 1999}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 420 rows and 18 columns.
}
\source{
\url{www.cde.ca.gov}
}
\usage{
caschool
}
\description{
@format {
\item{observation_n~r}{float, \%9.0g}
\item{dist_cod}{float, \%9.0g}
\item{county}{str18, \%18s}
\item{district}{str53, \%18s}
...
}
}
\keyword{datasets}
