\name{Caschool}
\docType{data}
\alias{Caschool}
\title{The California test score data set}
\description{
  a cross-section from 1998-1999

 \emph{number of observations} :  420

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Caschool)}
\format{A dataframe containing :
  \describe{
    \item{ID}{observation number}
    \item{distcod}{disctric code}
    \item{county}{county}
    \item{district}{district}
    \item{grspan}{grade span of district}
    \item{enrltot}{total enrollment}
    \item{teachers}{number of teachers}
    \item{calwpct}{percent qualifying for CalWorks}
    \item{mealpct}{percent qualifying for reduced-price lunch}
    \item{computer}{number of computers}
    \item{testscr}{average test score (read.scr+math.scr)/2}
    \item{compstu}{computer per student}
    \item{expnstu}{expenditure per student}
    \item{str}{student teacher ratio}
    \item{avginc}{district average income}
    \item{elpct}{percent of English learners}
    \item{readscr}{average reading score}
    \item{mathscr}{average math score}
  }
}
\source{
    California Department of Education \url{www.cde.ca.gov}.
}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers.
}
\keyword{datasets}